refactor: abstracted components and put them in a separate files + renamed component
- Section -> StepSection
- Abstracted primary button in a separate file
- created generic type file
