+++ apps/examples/nextjs/src/components/types/generic.ts
@@ -1,0 +1,7 @@
+import {ReactNode} from "react";
+
+export interface ButtonProps {
+    disabled?: boolean
+    children: ReactNode
+}
+
